[RFR] Code reorganization and production optimizations - [x] Removed `common-client` in favor of `isomorphic` directory - [x] Each project now have its own `.eslintrc` - [x] Added default `staging` configuration - [x] Optimized webpack build for production (Made sure NODE_ENV was set to get react and redux in production mode)
