+++ config/default.js
@@ -80,7 +80,6 @@ module.exports = {
         },
         frontend: {
             api_url: `${apiUrl}/api`,
-            enableDevTools: true,
         },
     },
     babel_ignore: /node_modules\/(?!admin-config|fakerest)/,
