+++ package.json
@@ -91,12 +91,13 @@
     "db-migrate": "0.9.25",
     "db-migrate-pg": "0.1.10",
     "enzyme": "2.4.1",
-    "eslint": "3.4.0",
-    "eslint-config-airbnb": "10.0.1",
-    "eslint-plugin-import": "1.14.0",
-    "eslint-plugin-jsx-a11y": "2.2.1",
+    "eslint": "3.7.0",
+    "eslint-config-airbnb": "12.0.0",
+    "eslint-config-airbnb-base": "8.0.0",
+    "eslint-plugin-import": "2.0.0",
+    "eslint-plugin-jsx-a11y": "2.2.2",
     "eslint-plugin-mocha": "4.5.1",
-    "eslint-plugin-react": "6.2.0",
+    "eslint-plugin-react": "6.3.0",
     "expect": "1.20.2",
     "expose-loader": "0.7.1",
     "extract-text-webpack-plugin": "1.0.1",
@@ -110,7 +111,7 @@
     "mocha": "3.0.2",
     "node-sass": "3.9.0",
     "nodium": "0.1.0",
-    "pm2": "1.1.3",
+    "pm2": "2.0.15",
     "react-addons-test-utils": "15.3.1",
     "redux-mock-store": "1.1.4",
     "sass-loader": "4.0.1",
