--- src/admin/css/login.scss
@@ -1,44 +1,0 @@
-@import "~bootstrap/scss/bootstrap";
-@import "compass/css3";
-
-body {
-    background-color: #eee;
-    padding-bottom: 40px;
-    padding-top: 40px;
-}
-.form-signin {
-    margin: 0 auto;
-    max-width: 330px;
-    padding: 15px;
-
-    #loginError {
-        display: none;
-    }
-    .form-signin-heading,
-    .checkbox {
-        margin-bottom: 10px;
-    }
-    .checkbox {
-        font-weight: normal;
-    }
-    .form-control {
-        box-sizing: border-box;
-        font-size: 16px;
-        height: auto;
-        padding: 10px;
-        position: relative;
-    }
-    .form-control:focus {
-        z-index: 2;
-    }
-    input[type="email"] {
-        border-bottom-left-radius: 0;
-        border-bottom-right-radius: 0;
-        margin-bottom: -1px;
-    }
-    input[type="password"] {
-        border-top-left-radius: 0;
-        border-top-right-radius: 0;
-        margin-bottom: 10px;
-    }
-}
