--- src/admin/login.html
@@ -1,24 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="utf-8" />
-        <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1, user-scalable=0" />
-        <meta http-equiv="x-ua-compatible" content="ie=edge" />
-        <title>New App - Sign-in</title>
-    </head>
-    <body>
-        <div class="container">
-            <form class="form-signin" id="loginForm">
-                <h2 class="form-signin-heading">Sign in</h2>
-                <div class="alert alert-danger" id="loginError" role="alert">
-                    <strong>Login Error</strong> Please try again
-                </div>
-                <label for="inputEmail" class="sr-only">Email address</label>
-                <input type="email" id="inputEmail" class="form-control" placeholder="Email address" required autofocus>
-                <label for="inputPassword" class="sr-only">Password</label>
-                <input type="password" id="inputPassword" class="form-control" placeholder="Password" required>
-                <button class="btn btn-lg btn-primary btn-block" type="submit" id="submitButton">Sign in</button>
-            </form>
-        </div>
-    </body>
-</html>
