+++ src/frontend/.eslintrc
@@ -1,0 +1,17 @@
+{
+    "parser": "babel-eslint",
+    "env": {
+        "browser": true,
+        "mocha": true,
+        "node": true,
+    },
+    "extends": "airbnb",
+    "rules": {
+        "indent": ["error", 4],
+        "max-len": ["error", {"code": 120, "tabWidth": 4, "ignoreUrls": true}],
+        "react/jsx-filename-extension": ["warn", { "extensions": [".js", ".jsx"] }],
+        "react/jsx-indent": ["error", 4],
+        "react/jsx-indent-props": ["error", 4],
+        "import/no-extraneous-dependencies": "off"
+    },
+}
