+++ src/frontend/js/order/api.js
@@ -1,5 +1,5 @@
 /* globals API_URL */
-import fetchFactory from '../../../common-client/fetch/fetch';
+import fetchFactory from '../../../isomorphic/fetch/fetch';
 
 export const fetchOrders = fetchFactory('orders');
 export const fetchOrder = fetchFactory('orders');
