+++ src/frontend/js/product/sagas.js
@@ -1,5 +1,5 @@
 import productActions, { productActionTypes } from './actions';
 import { fetchProduct, fetchProducts } from './api';
-import { createEntitySagas } from '../../../common-client/fetch/sagas';
+import { createEntitySagas } from '../../../isomorphic/fetch/sagas';
 
 export default createEntitySagas(productActionTypes, productActions, fetchProducts, fetchProduct);
