+++ src/isomorphic/fetch/createFetchActionTypes.js
