+++ src/isomorphic/fetch/createFetchActionTypes.spec.js
